added logic for logistic tast
